docs: draft a decision-making process
The decisions/README in the docs folder proposes a simple and manual
process to keep a record of decisions made along the way.

While still a draft, I think it is worth committing to capture the
evolution of that process. It will make more sense when combined with
future commits.