docs: shorten numbering of decision records
I think three digits (up to a thousand) worth of decision-records is a
bit excessive at this point. I chose to shorten the allocation to only
two digits (a hundred) knowing that by the time more decisions are made
and recorded, the pool can be extended.

To extend the numbering scheme back by a factor of ten, simply rename
all files (and their links) by prefixing a zero character.

There may be more examples of file names floating around. Those can be
found by searching for "00-" or "01-".