github: remove unused Hugo dependencies in the publishing workflow
The default Hugo template installs dart-sass and npm tools. Currently,
we use neither, so there's no harm in that. When we finally expand the
blog to require these advanced capabilities, the tools can be easily
reimplemented.

I chose to keep them until this point to preserve their existance in the
commit history.