github: publish blog using the latest Hugo release
Keeping CI on the latest release confronts us maintainers with breaking
changes earlier. I find it crucial to upgrade to the latest stable
release as soon as possible, otherwise upgrades have a higher risk
because the cost of upgrading grows as the version gap widens.

It's mostly a matter of personal ideology and style. I like to keep all
tools updated.